# backend/endoscope/docs/api.yaml
openapi: 3.0.3
info:
  title: API for endoscope
  description: API for endoscope
  contact:
    name: endoscope
    email: endoscope@localhost
  version: 1.0.0
...
paths:
  /auth/signup:
    post:
      tags:
        - auth
      summary: Register a new user
      description: Register a new user
      operationId: registerUser
      requestBody:
        description: User registration data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEntity'
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user and generate JWT token
      description: Authenticate user with email and password, and generate JWT token
      operationId: loginUser
      requestBody:
        description: User login credentials
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /auth/generateBackUp:
    post:
      tags:
        - auth
      summary: Generate backup of the database
      description: Generate backup of the database
      operationId: generateBackUp
      responses:
        '200':
          description: Backup generated successfully
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /auth/uploadBackup:
    post:
      tags:
        - auth
      summary: Upload backup of the database
      description: Upload backup of the database
      operationId: uploadBackUp
      responses:
        '200':
          description: Backup upload successfully
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

  #############################
  # Directory API             #
  #############################
...
components:
  schemas:
    UserEntity:
      type: object
      properties:
        userId:
          type: integer
          description: The user's ID
        email:
          type: string
          format: email
          description: The user's email address
        password:
          type: string
          format: password
          description: The user's password
        salutation:
          type: string
          description: The user's salutation (e.g., Mr., Ms., Dr.)
        gender:
          type: string
          description: The user's gender
        fullName:
          type: string
          description: The user's full name
        age:
          type: string
          description: The user's age
        dateOfBirth:
          type: string
          format: date
          description: The user's date of birth
        countryOfOrigin:
          type: string
          description: The user's country of origin
        workLocation:
          type: string
          description: The user's work location
        workPlaceSetting:
          type: string
          description: The user's workplace setting
        medicalId:
          type: string
          description: The user's medical ID
        medicalSpeciality:
          type: string
          description: The user's medical specialty
        education:
          type: string
          description: The user's education
        role:
          type: string
          description: The user's role
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
          description: The user's token
        password:
          type: string
          format: password
          description: The user's password
    UserLoginResponse:
      type: object
      properties:
        token:
          type: string
          format: string
          description: The user's token
        expiration:
          type: long
          description: The expiration time of the token
        email:
          type: string
          format: string
          description: The user's email
        name:
          type: string
          format: string
          description: The user's name
        role:
          type: string
          format: string
          description: The user's role
...
    UnauthorizedError:
      type: object
      properties:
        message:
          type: string
          description: Error message
    BadRequestError:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
    InternalServerError:
      type: object
      properties:
        message:
          type: string
          description: The error message
        code:
          type: integer
          description: The error code